Module memory_magic::raw

Functions for allocating and deallocating virtual memory.

Structs

FileOptions

Options for opening a file mapping.

FilePermissions

Permissions for file mapping.

Length

A length of a view into an object.

Object

An object that can be mapped to virtual memory.

Offset

An offset into an object where a view may begin.

View

A view of an object.

ViewMut

A mutable view of an object.

Enums

ReadPermissions

Permissions for a particular view of an object.

WritePermissions

Permissions for a particular mutable view of an object.

Functions

map

Map a view of an object to memory.

map_multiple

Map views of objects contiguously to memory.

map_multiple_mut

Map mutable views of objects contiguously to memory.

map_mut

Map a mutable view of an object to memory.

unmap⚠

Unmap a memory map.